Split out differnt data structures
Also implement naive 'predict probability of a city' function. *Does
not* take into account the fact that the card has already been
reshuffled from an epidemic
