+++ pandemic/cities_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 )
 
-func TestNeighbors(t *testing.T) {
+func TestLoadFromJSON(t *testing.T) {
 
 	// Do IO
 	filename, _ := filepath.Abs("../data/cities.json")
@@ -22,6 +22,7 @@ func TestNeighbors(t *testing.T) {
 		fmt.Errorf("Decode city: %v", err)
 	}
 
-	fmt.Println(AllCitiesWithDisease(c.Cities, Blue))
+}
 
+func TestProbabilityOfCard(t *testing.T) {
 }
